requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.1
django-po2xls==0.4.0
django-rosetta==0.7.13
django-xls2po==0.3.0
et-xmlfile==1.0.1
geoip2==2.9.0
idna==2.8
importlib-metadata==1.5.0
jdcal==1.4.1
maxminddb==1.4.1
microsofttranslator==0.8
numpy==1.16.3
openpyxl==2.6.2
pandas==0.24.2
path==13.1.0
Pillow==6.0.0
polib==1.1.0
psycopg2==2.8.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
utils==1.0.1
wx-sdk==1.0.5
xlrd==1.0.0
xlwt==1.2.0
zipp==2.2.0
